Use PSBT input pairs for input contributions
Previously an "input" consisted of a (OutPoint, TxOut) tuple, which
limits the caller to only contributing witness inputs. Instead, redefine
an input as a (psbt::Input, TxIn) tuple, which is more flexible as it
allows for any necessary information to be included in the PSBT input.

Two PSBT input fields we are particularly interested in are
`non_witness_utxo` and `redeem_script`, which enable legacy inputs and
nested segwit (p2wpkh-in-p2sh) respectively.

`contribute_witness_inputs` is renamed to `contribute_inputs`.