Fix weight estimations for nested segwit inputs
To determine whether an input script uses nested segwit
(p2wpkh-in-p2sh), we need the redeem script that unlocks it. This can be
obtained from either the input signature (if the input is signed, e.g.
the sender inputs ), or the redeem_script field on the PSBT input which
can be populated by the wallet that owns it (e.g. the receiver inputs).

When making weight predictions for a P2SH input, we now check the
final_script_sig first and fallback to the redeem_script if there is no
signature. Additionally, the order of operations in `finalize_proposal`
is updated such that we don't clear the sender's signatures until after
the weight predictions/fee estimations are completed.