Specify bindgen binary path based on the OS when using cargo run
The uniffi_bindgen requires that the source bindary be a certain
filepath format. This uses the std lib to determine the OS type and
creates a simple string to use as the path. We currently on actually
support Linux and MacOS but I thought it seemed easy enought to add
future support for Windows here.